+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v2.9.3
 
+* `TerminableThread.start` now returns self